// projectpineers/projectpioneers/src/app/models/collaborator.ts
export interface Collaborator {
        globalGroupId: number;
        firstName: string ;
        lastName: string;
        email: string;
        officeBase: string;
        birthday: Date;
        status: string;
        classification: string;
        groupGrade: string;
        localGrade: string;
        puCode: string;
        jobTitle: string;
        typeContrat: string;
        projectName: string;
        joiningDate: Date;
};
      